Add files via upload
We propose a framework for consistently applying the 2018 periodontal status classification scheme to epidemiological surveys (Application of the 2018 periodontal status Classification to Epidemiological Survey data, ACES). We specified data requirements and workflows for either completed or planned epidemiological surveys, utilizing commonly collected measures of periodontal status [clinical attachment levels (CAL), probing depths (PD), bleeding on probing (BoP)], as well as additional necessary variables for the implementation of the 2018 periodontal status classification (tooth loss due to periodontitis and complexity factors). Following detailed instructions and flowcharts, survey participants are classified as having periodontal health, gingivitis, or periodontitis. Rates of edentulism must also be reported. In cases of periodontitis, instructions on how to compute Stage and Extent are provided. Assessment of Grade can be derived from CAL measurements (or from radiographic alveolar bone loss data) in relation to root length and the participant’s age. ACES is a framework to be used in epidemiological studies of periodontal status that (i) have been completed, and in which Stage and Grade according to the 2018 classification are inferred retroactively, or (ii) are being planned. Consistent use of the proposed comprehensive approach will facilitate the comparability of periodontitis prevalence estimates across studies.